requirement.txt
mistralai
python-dotenv
pygame
Unidecode